Add maven execution id, remove execution phase and remove reportSets from pom.xml
The maven execution id is not necessary, but it looks good and it will be displayed while running. The default
execution phase of the plugin goals used by this project are Ok, so there is no need to specify them.
The reportSet for maven-javadoc-plugin is not necessary as the default goal is Ok.